requirements.txt
flask
azure-storage-blob
python-dotenv
